# api/openapi.yaml
openapi: 3.2.0
info:
  description: Per-task staging environments. Tasks belong to projects and are served
    at /{project}/{task}/. Proxied task traffic is public; this API is not.
  title: Boreas API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      description: The token is returned once and expires in 30 days.
      operationId: post-auth-login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Exchange credentials for a token
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      operationId: post-auth-logout
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Revoke the current token
      tags:
      - auth
  /api/v1/auth/me:
    get:
      operationId: get-auth-me
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Current user
      tags:
      - auth
  /api/v1/auth/tokens:
    get:
      description: Metadata only; tokens and hashes are never returned.
      operationId: get-auth-tokens
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APITokensResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List your API tokens
      tags:
      - auth
    post:
      description: The token is returned once. Validity must not exceed 90 days. Requires
        a login session, not another API token.
      operationId: post-auth-tokens
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPITokenRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPITokenResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create an API token
      tags:
      - auth
  /api/v1/auth/tokens/{id}:
    delete:
      description: Requires a login session. Users revoke only their own tokens.
      operationId: delete-auth-tokens-id
      parameters:
      - in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Revoke one of your API tokens
      tags:
      - auth
...
  /api/v1/notifications:
    get:
      description: Newest first, from every project and task the caller reaches. To
        page, pass the id of the last notification received as before.
      operationId: get-notifications
      parameters:
      - in: query
        name: limit
        schema:
          default: 50
          maximum: 200
          minimum: 1
          type: integer
      - in: query
        name: before
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List notifications across projects
      tags:
      - notifications
  /api/v1/notifications/seen:
    post:
      description: Seen is tracked per user. Idempotent; ids outside the caller's
        visibility are skipped.
      operationId: post-notifications-seen
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkSeenRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Mark notifications seen
      tags:
      - notifications
  /api/v1/projects:
    get:
      description: Administrators see every project, others their memberships and
        grants. my_role is the caller's role, per project and per task, where a grant
        can raise it; last_deploy is the newest deploy.
      operationId: get-projects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List reachable projects with their tasks
      tags:
      - projects
    post:
      description: The creator becomes owner. The slugs api, health, metrics, static
        and admin are reserved. default_image, default_port and default_env only prefill
        the task creation form; task creation never applies them on its own.
      operationId: post-projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create a project
      tags:
      - projects
...
  /api/v1/projects/{project}/members:
    get:
      operationId: get-projects-project-members
      parameters:
      - in: path
        name: project
        required: true
        schema:
          example: demo
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembersResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List project members
      tags:
      - projects
    post:
      operationId: post-projects-project-members
      parameters:
      - in: path
        name: project
        required: true
        schema:
          example: demo
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Add or promote a member
      tags:
      - projects
...
  /api/v1/public/projects:
    get:
      description: Every project with its tasks, reduced to what the public directory
        pages show. Image, port, environment and container identifiers are never included.
      operationId: get-public-projects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProjectsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Public environment directory
      tags:
      - system
...
  /api/v1/stats:
    get:
      operationId: get-stats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStats'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Service statistics
      tags:
      - system
...
components:
  schemas:
...
    APITokensResponse:
      properties:
        api_tokens:
          items:
            $ref: '#/components/schemas/APIToken'
          nullable: true
          type: array
        total:
          type: integer
      required:
      - api_tokens
      - total
      type: object
...
    CreateAPITokenResponse:
      properties:
        api_token:
          $ref: '#/components/schemas/APIToken'
        token:
          format: password
          type: string
      required:
      - token
      - api_token
      type: object
...
    ErrorResponse:
      properties:
        error:
          example: not found
          type: string
      required:
      - error
      type: object
...
    LoginRequest:
      properties:
        password:
          format: password
          type: string
        username:
          example: admin
          type: string
      required:
      - username
      - password
      type: object
    LoginResponse:
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
      - token
      - user
      type: object
...
    SuccessResponse:
      properties:
        success:
          example: true
          type: boolean
      required:
      - success
      type: object
...
    UserResponse:
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
      - user
      type: object
...
  securitySchemes:
    bearerAuth:
      description: 'Login session or API token, sent as: Authorization: Bearer <token>'
      scheme: bearer
      type: http